+++ backend/src/main/java/codesquad/team01/issuetracker/label/repository/LabelQueryRepository.java
@@ -5,6 +5,6 @@ import java.util.List;
 import codesquad.team01.issuetracker.label.dto.LabelDto;
 
 public interface LabelQueryRepository {
-	List<LabelDto.IssueLabelRow> findLabelsByIssueIds(List<Long> issueIds);
+	List<LabelDto.IssueLabelRow> findLabelsByIssueIds(List<Integer> issueIds);
 
 }
