+++ backend/src/main/java/codesquad/team01/issuetracker/user/repository/UserQueryRepository.java
@@ -6,5 +6,5 @@ import codesquad.team01.issuetracker.user.dto.UserDto;
 
 public interface UserQueryRepository {
 
-	List<UserDto.IssueAssigneeRow> findAssigneesByIssueIds(List<Long> issueIds);
+	List<UserDto.IssueAssigneeRow> findAssigneesByIssueIds(List<Integer> issueIds);
 }
